pyproject.toml
[build-system]
requires      = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "ragebot-mcp"
version     = "1.0.0"
description = "Intelligent CLI-based Project Context Engine with MCP server support"
readme      = "README.md"
requires-python = ">=3.10"
license     = { text = "MIT" }
authors     = [{ name = "RageBot Team" }]
keywords    = ["cli", "ai", "mcp", "gemini", "grok", "developer-tools", "code-context"]

# ── Core dependencies (installed with: pip install ragebot-mcp) ──────────────
dependencies = [
    # CLI framework
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    # Embeddings - required for semantic search (downloads ~90 MB model on first use)
    "sentence-transformers>=2.7.0",
    # Token counting
    "tiktoken>=0.5.0",
    # Secure API key storage in OS keyring
    "keyring>=24.0.0",
    # .env file loading
    "python-dotenv>=1.0.0",
]

# ── Optional extras ───────────────────────────────────────────────────────────
[project.optional-dependencies]

# LLM providers
gemini = [
    "google-generativeai>=0.7.0",
]
grok = [
    # Grok uses an OpenAI-compatible API - the openai SDK is all that's needed
    "openai>=1.20.0",
]

# Faster vector search (recommended for large codebases)
vector = [
    "faiss-cpu>=1.7.4",
]

# Document parsing (PDF, DOCX)
docs_extra = [
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "pdfplumber>=0.10.0",
]

# HTTP/SSE transport for remote MCP deployments
sse = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
]

# File-system watcher for `rage watch`
watch = [
    "watchdog>=3.0.0",
]

# Install everything at once
full = [
    "ragebot-mcp[gemini,grok,vector,docs_extra,sse,watch]",
]

# Development tools
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

# ── CLI entry points ──────────────────────────────────────────────────────────
[project.scripts]
rage                = "ragebot.cli:main"
ragebot-mcp-server  = "ragebot.mcp.server:main"

[project.entry-points."mcp.servers"]
ragebot = "ragebot.mcp.server:main"

# ── Package discovery ─────────────────────────────────────────────────────────
[tool.setuptools.packages.find]
where   = ["."]
include = ["ragebot*"]

# ── Tool config ───────────────────────────────────────────────────────────────
[tool.black]
line-length    = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length    = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version         = "3.10"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
